+++ src/main/java/com/codessquad/qna/web/controller/AnswerController.java
@@ -1,14 +1,6 @@
 package com.codessquad.qna.web.controller;
 
-import com.codessquad.qna.web.domain.answer.Answer;
-import com.codessquad.qna.web.domain.answer.AnswerRepository;
-import com.codessquad.qna.web.domain.question.Question;
-import com.codessquad.qna.web.domain.question.QuestionRepository;
-import com.codessquad.qna.web.domain.user.User;
-import com.codessquad.qna.web.exception.AnswerNotFoundException;
-import com.codessquad.qna.web.exception.CRUDAuthenticationException;
-import com.codessquad.qna.web.exception.QuestionNotFoundException;
-import com.codessquad.qna.web.utils.SessionUtils;
+import com.codessquad.qna.web.service.AnswerService;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -21,44 +13,21 @@ import javax.servlet.http.HttpSession;
 @RequestMapping("/questions/{questionId}/answers")
 public class AnswerController {
 
-    private final AnswerRepository answerRepository;
-    private final QuestionRepository questionRepository;
+    private final AnswerService answerService;
 
-    public AnswerController(AnswerRepository answerRepository, QuestionRepository questionRepository) {
-        this.answerRepository = answerRepository;
-        this.questionRepository = questionRepository;
+    public AnswerController(AnswerService answerService) {
+        this.answerService = answerService;
     }
 
     @PostMapping()
-    public String create(@PathVariable Long questionId, String contents, HttpSession session) {
-        User loginUser = SessionUtils.getLoginUser(session);
-
-        Question question = getQuestionById(questionId);
-        Answer answer = Answer.toEntity(loginUser, question, contents);
-        answerRepository.save(answer);
-
+    public String create(@PathVariable long questionId, String contents, HttpSession session) {
+        answerService.create(questionId, contents, session);
         return "redirect:/questions/" + questionId;
     }
 
     @DeleteMapping("/{id}")
-    public String delete(@PathVariable Long questionId, @PathVariable Long id, HttpSession session) {
-        User loginUser = SessionUtils.getLoginUser(session);
-        Answer answer = answerRepository.findById(id)
-                .orElseThrow(() -> new AnswerNotFoundException("Cannot found answer number " + id));
-
-        if (!answer.isMatchingWriter(loginUser)) {
-            throw new CRUDAuthenticationException("Only writer can delete this answer post!");
-        }
-
-        answerRepository.delete(answer);
-
+    public String delete(@PathVariable long questionId, @PathVariable long id, HttpSession session) {
+        answerService.delete(questionId, id, session);
         return "redirect:/questions/" + questionId;
     }
-
-    private Question getQuestionById(Long id) {
-        return questionRepository.findById(id)
-                .orElseThrow(() -> new QuestionNotFoundException("Cannot found question number " + id));
-    }
-
-
 }
